src/datamigration/scripts/teradata/agent_controller/requirements.txt
google-cloud-pubsub
PyYAML